client/requirements.txt
pydivert
websockets
pystray
Pillow
